# optimization around get for raw usage